+++ packages/card/src/components/StatisticCard/demos/basic.tsx
@@ -1,16 +1,17 @@
 import { EllipsisOutlined, RightOutlined } from '@ant-design/icons';
 import { StatisticCard } from '@ant-design/pro-components';
-import { Space } from 'antd';
+import { Space, theme } from 'antd';
 
 const { Statistic } = StatisticCard;
 
 export default () => {
+  const { token } = theme.useToken();
   return (
     <StatisticCard
       title={
         <Space>
           <span>部门一</span>
-          <RightOutlined style={{ color: 'rgba(0,0,0,0.65)' }} />
+          <RightOutlined style={{ color: token.colorTextHeading }} />
         </Space>
       }
       extra={<EllipsisOutlined />}