+++ packages/field/src/components/Code/index.tsx
@@ -3,6 +3,7 @@ import React from 'react';
 import type { ProFieldFC } from '../../index';
 
 // 兼容代码-----------
+import { proTheme } from '@ant-design/pro-provider';
 import 'antd/lib/input/style';
 //----------------------
 
@@ -33,6 +34,7 @@ const FieldCode: ProFieldFC<{
   ref,
 ) => {
   const code = languageFormat(text, language);
+  const { token } = proTheme.useToken();
   if (mode === 'read') {
     const dom = (
       <pre
@@ -43,7 +45,9 @@ const FieldCode: ProFieldFC<{
           overflow: 'auto',
           fontSize: '85%',
           lineHeight: 1.45,
-          backgroundColor: '#f6f8fa',
+          color: token.colorTextSecondary,
+          fontFamily: token.fontFamilyCode,
+          backgroundColor: 'rgba(150, 150, 150, 0.1)',
           borderRadius: 3,
           width: 'min-content',
           ...fieldProps.style,