+++ packages/field/src/components/Select/SearchSelect/index.tsx
@@ -2,7 +2,7 @@ import type { RequestOptionsType } from '@ant-design/pro-utils';
 import type { SelectProps } from 'antd';
 import { ConfigProvider, Select } from 'antd';
 
-import type { LabeledValue } from 'antd/lib/select';
+import type { DefaultOptionType, LabeledValue } from 'antd/lib/select';
 import classNames from 'classnames';
 import React, {
   useContext,
@@ -11,8 +11,6 @@ import React, {
   useRef,
   useState,
 } from 'react';
-
-const { Option, OptGroup } = Select;
 
 // 支持 key, value, label，兼容 UserSearch 中只填写了 key 的情况。
 export type KeyLabel = Partial<LabeledValue> & RequestOptionsType;
@@ -168,12 +166,14 @@ const SearchSelect = <T,>(props: SearchSelectProps<T[]>, ref: any) => {
     return [];
   };
 
-  const renderOptions = (mapOptions: RequestOptionsType[]) => {
-    return mapOptions.map((item) => {
+  const genOptions = (
+    mapOptions: RequestOptionsType[],
+  ): DefaultOptionType[] => {
+    return mapOptions.map((item, index) => {
       const {
-        disabled: itemDisable,
         className: itemClassName,
         optionType,
+        ...resetItem
       } = item as RequestOptionsType;
 
       const label = item[labelPropsName];
@@ -181,26 +181,24 @@ const SearchSelect = <T,>(props: SearchSelectProps<T[]>, ref: any) => {
       const itemOptions = item[optionsPropsName] ?? [];
 
       if (optionType === 'optGroup' || item.options) {
-        return (
-          <OptGroup key={value} label={label}>
-            {renderOptions(itemOptions)}
-          </OptGroup>
-        );
+        return {
+          ...resetItem,
+          label: label,
+          title: label,
+          key: value ?? label?.toString(),
+          children: genOptions(itemOptions),
+        } as DefaultOptionType;
       }
 
-      return (
-        <Option
-          {...item}
-          value={value!}
-          key={value || label?.toString()}
-          disabled={itemDisable}
-          data-item={item}
-          className={`${prefixCls}-option ${itemClassName || ''}`}
-          label={label}
-        >
-          {optionItemRender?.(item as any) || label}
-        </Option>
-      );
+      return {
+        ...resetItem,
+        title: label,
+        value: value ?? index,
+        key: value ?? label?.toString(),
+        'data-item': item,
+        className: `${prefixCls}-option ${itemClassName || ''}`.trim(),
+        label: optionItemRender?.(item as any) || label,
+      } as DefaultOptionType;
     });
   };
   return (
@@ -215,18 +213,6 @@ const SearchSelect = <T,>(props: SearchSelectProps<T[]>, ref: any) => {
       searchValue={searchValue}
       optionFilterProp={optionFilterProp}
       optionLabelProp={optionLabelProp}
-      filterOption={(inputValue, option) => {
-        return !!(
-          option?.label
-            ?.toString()
-            .toLowerCase()
-            .includes(inputValue.toLowerCase()) ||
-          option?.value
-            ?.toString()
-            .toLowerCase()
-            .includes(inputValue.toLowerCase())
-        );
-      }} // 这里使用pro-components的过滤逻辑
       onClear={() => {
         onClear?.();
         fetchData(undefined);
@@ -235,6 +221,35 @@ const SearchSelect = <T,>(props: SearchSelectProps<T[]>, ref: any) => {
         }
       }}
       {...restProps}
+      filterOption={
+        restProps.filterOption == false
+          ? false
+          : (inputValue, option) => {
+              if (
+                restProps.filterOption &&
+                typeof restProps.filterOption === 'function'
+              ) {
+                return restProps.filterOption(inputValue, {
+                  ...option,
+                  label: option?.title,
+                });
+              }
+              return !!(
+                option?.title
+                  ?.toString()
+                  .toLowerCase()
+                  .includes(inputValue.toLowerCase()) ||
+                option?.label
+                  ?.toString()
+                  .toLowerCase()
+                  .includes(inputValue.toLowerCase()) ||
+                option?.value
+                  ?.toString()
+                  .toLowerCase()
+                  .includes(inputValue.toLowerCase())
+              );
+            }
+      } // 这里使用pro-components的过滤逻辑
       onSearch={
         showSearch
           ? (value) => {
@@ -283,9 +298,8 @@ const SearchSelect = <T,>(props: SearchSelectProps<T[]>, ref: any) => {
         }
         onFocus?.(e);
       }}
-    >
-      {renderOptions(options || [])}
-    </Select>
+      options={genOptions(options || [])}
+    />
   );
 };