+++ packages/layout/src/components/SiderMenu/SiderMenu.tsx
@@ -1,6 +1,6 @@
 import type { GenerateStyle } from '@ant-design/pro-provider';
 import { ProProvider } from '@ant-design/pro-provider';
-import type { AvatarProps, SiderProps} from 'antd';
+import type { AvatarProps, SiderProps } from 'antd';
 import { Avatar, Layout, Menu, Space } from 'antd';
 import type { ItemType } from 'antd/lib/menu/hooks/useItems';
 import classNames from 'classnames';