+++ packages/list/src/Item.tsx
@@ -1,6 +1,6 @@
 import { RightOutlined } from '@ant-design/icons';
-import type { ProCardProps } from '@ant-design/pro-card';
-import ProCard from '@ant-design/pro-card';
+import type { CheckCardProps } from '@ant-design/pro-card';
+import { CheckCard } from '@ant-design/pro-card';
 import { ProProvider } from '@ant-design/pro-provider';
 import { ConfigProvider, List, Skeleton } from 'antd';
 
@@ -91,7 +91,7 @@ export type ItemProps<RecordType> = {
   type?: 'new' | 'top' | 'inline' | 'subheader';
   isEditable: boolean;
   recordKey: string | number | undefined;
-  cardProps?: ProCardProps;
+  cardProps?: CheckCardProps;
   record: RecordType;
   onRow?: GetComponentProps<RecordType>;
   onItem?: GetComponentProps<RecordType>;
@@ -207,22 +207,32 @@ function ProListItem<RecordType>(props: ItemProps<RecordType>) {
     return [
       <div
         key="action"
-        className={`${className}-actions ${hashId}`.trim()}
+        className={`${defaultClassName}-actions ${hashId}`.trim()}
         onClick={(e) => e.stopPropagation()}
       >
         {actions}
       </div>,
     ];
-  }, [actions, cardActionProps, className, hashId]);
+  }, [actions, cardActionProps, defaultClassName, hashId]);
 
   const titleDom =
     title || subTitle ? (
-      <div className={`${className}-header-title ${hashId}`.trim()}>
+      <div className={`${defaultClassName}-header-container ${hashId}`.trim()}>
         {title && (
-          <div className={`${className}-title ${hashId}`.trim()}>{title}</div>
+          <div
+            className={classNames(`${defaultClassName}-title`, hashId, {
+              [`${defaultClassName}-title-editable`]: isEditable,
+            })}
+          >
+            {title}
+          </div>
         )}
         {subTitle && (
-          <div className={`${className}-subTitle ${hashId}`.trim()}>
+          <div
+            className={classNames(`${defaultClassName}-subTitle`, hashId, {
+              [`${defaultClassName}-subTitle-editable`]: isEditable,
+            })}
+          >
             {subTitle}
           </div>
         )}
@@ -248,8 +258,8 @@ function ProListItem<RecordType>(props: ItemProps<RecordType>) {
     ) : null;
 
   const rowClassName = classNames(hashId, {
-    [`${className}-item-has-checkbox`]: checkbox,
-    [`${className}-item-has-avatar`]: avatar,
+    [`${defaultClassName}-item-has-checkbox`]: checkbox,
+    [`${defaultClassName}-item-has-avatar`]: avatar,
     [className]: className,
   });
   const cardTitleDom = useMemo(() => {
@@ -270,6 +280,7 @@ function ProListItem<RecordType>(props: ItemProps<RecordType>) {
     return null;
   }, [avatar, getPrefixCls, hashId, title]);
 
+  const itemProps = onItem?.(record, index);
   const defaultDom = !cardProps ? (
     <List.Item
       className={classNames(rowClassName, hashId, {
@@ -279,7 +290,7 @@ function ProListItem<RecordType>(props: ItemProps<RecordType>) {
       actions={extraDom}
       extra={!!extra && <div className={extraClassName}>{extra}</div>}
       {...onRow?.(record, index)}
-      {...onItem?.(record, index)}
+      {...itemProps}
       onClick={(e) => {
         onRow?.(record, index)?.onClick?.(e);
         onItem?.(record, index)?.onClick?.(e);
@@ -328,10 +339,11 @@ function ProListItem<RecordType>(props: ItemProps<RecordType>) {
       </Skeleton>
     </List.Item>
   ) : (
-    <ProCard
+    <CheckCard
       bordered
-      loading={loading}
-      hoverable
+      style={{
+        width: '100%',
+      }}
       {...cardProps}
       title={cardTitleDom}
       subTitle={subTitle}
@@ -341,7 +353,11 @@ function ProListItem<RecordType>(props: ItemProps<RecordType>) {
         padding: 24,
         ...cardProps.bodyStyle,
       }}
-      {...onItem?.(record, index)}
+      {...(itemProps as CheckCardProps)}
+      onClick={(e: any) => {
+        cardProps?.onClick?.(e);
+        itemProps?.onClick?.(e);
+      }}
     >
       <Skeleton avatar title={false} loading={loading} active>
         <div className={`${className}-header ${hashId}`.trim()}>
@@ -349,7 +365,7 @@ function ProListItem<RecordType>(props: ItemProps<RecordType>) {
           {content}
         </div>
       </Skeleton>
-    </ProCard>
+    </CheckCard>
   );
 
   if (!cardProps) {