+++ packages/list/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file. See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.5.5](https://github.com/ant-design/pro-components/compare/@ant-design/pro-list@2.5.4...@ant-design/pro-list@2.5.5) (2023-07-14)
+
+**Note:** Version bump only for package @ant-design/pro-list
+
 ## [2.5.4](https://github.com/ant-design/pro-components/compare/@ant-design/pro-list@2.5.3...@ant-design/pro-list@2.5.4) (2023-07-04)
 
 **Note:** Version bump only for package @ant-design/pro-list